requirements.txt
psycopg2-binary
python-dotenv~=1.0.0
websocket-client~=1.7.0
sqlalchemy~=2.0.23
python-binance
matplotlib~=3.8.2
numpy~=1.26.2
pandas~=2.1.3
ta~=0.11.0
requests
flask
flask_bootstrap